Switch again to using sqlalchemy and alembic
This commit removes flask-migrate and flask-sqlalchemy.

The intent of these 2 packages were meant to simplify running database
migrations and using sqlalchemy within a flask context, but they both ended up
just confusing things a bit more. It was overall simplier to user sqlalchemy and
alembic straight as opposed to these wrapper packages.

further reading:
https://towardsdatascience.com/use-flask-and-sqlalchemy-not-flask-sqlalchemy-5a64fafe22a4

This article shows that flask-sqlalchemy is not needed and actually makes it
more difficult to grab the db connection in a flask app context.

The removal of flask-sqlalchemy is what dictates the removal of flask-migrate,
as flask migrate relies on flask-sqlalchemy to operate.