Set up Database models
This commit sets up the main models for our database: Plant, Companion, Pest,
Schedule.

Plant being the "main" model, which has many companions, pests, and schedules

Companions are more of an association table between Plants and other plants. The
main addition being an `effect` column which describes how one plant affects
another. For example, a tomato plant will have marigolds as a companion and the
effect being that the marigold deters pests from the tomato. It is important to
note that the marigold plant does not have tomato as a companion as the tomato
has no effect on the growth of the marigold. So companions are considered a one
way relationship between the main and companionated plant.

Pests are as they seem, pests to plants. The thing that might need to be ironed
out later is the idea of pests as being generic vs how they effect certain
plants. We may need to add a similar table to companions in which it points to a
pest, but describes specifically how the main plant is affected by it. Currently
there is a secondary table called `pest_association` which accomplishes this
many-to-many relationship between plants and pests, but this will be ironed out
later.

Schedules is a bit generic right now in which the intended use is to hold USDA
hardiness zone planting information for when to plant indoor vs outdoor based on
which zone one lives in.